Change structure of shared links to random tests.
Previously, shared links to random tests depended on the backend because
each link was just a random UUID generated for each random test. This
change modifies the structure of the shared links such that all the
necessary information to recreate a random test is in the link itself.
This voids the need for any support from the backend/database.

For example, for a words-mode test, the structure of the path was
previously /:id where id was a randomly generated UUID. Now it would be
/words/:language/:length/:seed.